+++ NEWS.md
@@ -1,7 +1,7 @@
 # release 1.2
 
 This is a feature add release.
-- Adds bibliometric metadata and handling of multiple "vertiiical" sheets
+- Adds bibliometric metadata and handling of multiple "vertical" sheets
 - Adds autocreation of index file which summarises the metadata
 - Adds additional tests